Consolidate test_convert.py and rename
